refactor: move settings in `ui` directory
