+++ app/src/main/java/live/hms/android100ms/ui/meeting/MeetingViewModel.kt
@@ -8,7 +8,7 @@ import android.widget.Toast
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import live.hms.android100ms.model.RoomDetails
-import live.hms.android100ms.ui.home.settings.SettingsStore
+import live.hms.android100ms.ui.settings.SettingsStore
 import live.hms.android100ms.ui.meeting.chat.ChatMessage
 import live.hms.android100ms.util.*
 import live.hms.video.*
